+++ source/renderer/app/containers/wallet/Wallet.js
@@ -5,6 +5,7 @@ import MainLayout from '../MainLayout';
 import WalletWithNavigation from '../../components/wallet/layouts/WalletWithNavigation';
 import LoadingSpinner from '../../components/widgets/LoadingSpinner';
 import RestoreNotification from '../../components/notifications/RestoreNotification';
+import ChangeSpendingPasswordDialog from '../../components/wallet/settings/ChangeSpendingPasswordDialog';
 import { buildRoute } from '../../utils/routing';
 import { ROUTES } from '../../routes-config';
 import type { InjectedContainerProps } from '../../types/injectedPropsType';
@@ -50,7 +51,8 @@ export default class Wallet extends Component<Props> {
 
   render() {
     const { actions, stores } = this.props;
-    const { app, wallets } = stores;
+    const { app, wallets, uiDialogs } = stores;
+    const { isOpen: isDialogOpen } = uiDialogs;
     const { restartNode } = actions.networkStatus;
     const { active: activeWallet } = wallets;
 
@@ -70,7 +72,12 @@ export default class Wallet extends Component<Props> {
       recoveryPhraseVerificationStatus ===
         WalletRecoveryPhraseVerificationStatuses.NOTIFICATION &&
       !isIncentivizedTestnet;
-    const { isNotResponding, isRestoring } = activeWallet;
+    const {
+      isRestoring,
+      isLegacy,
+      isNotResponding,
+      hasPassword,
+    } = activeWallet;
 
     return (
       <MainLayout>
@@ -81,16 +88,23 @@ export default class Wallet extends Component<Props> {
         ) : null}
 
         <WalletWithNavigation
-          isActiveScreen={this.isActiveScreen}
-          onWalletNavItemClick={this.handleWalletNavItemClick}
           activeItem={app.currentPage}
-          isLegacy={activeWallet.isLegacy}
           hasNotification={hasNotification}
+          hasPassword={hasPassword}
+          isActiveScreen={this.isActiveScreen}
+          isLegacy={isLegacy}
           isNotResponding={isNotResponding}
+          isSetWalletPasswordDialogOpen={isDialogOpen(
+            ChangeSpendingPasswordDialog
+          )}
+          onOpenExternalLink={(url: string) => stores.app.openExternalLink(url)}
           onRestartNode={() => restartNode.trigger()}
-          onOpenExternalLink={(url: string) =>
-            this.props.stores.app.openExternalLink(url)
-          }
+          onSetWalletPassword={() => {
+            actions.dialogs.open.trigger({
+              dialog: ChangeSpendingPasswordDialog,
+            });
+          }}
+          onWalletNavItemClick={this.handleWalletNavItemClick}
         >
           {this.props.children}
         </WalletWithNavigation>