+++ source/renderer/app/containers/wallet/dialogs/settings/ChangeSpendingPasswordDialogContainer.js
@@ -8,7 +8,6 @@ import type { ActionsMap } from '../../../../actions/index';
 type Props = {
   stores: any | StoresMap,
   actions: any | ActionsMap,
-  forceSetPassword?: boolean,
 };
 
 @inject('actions', 'stores')
@@ -17,11 +16,10 @@ export default class ChangeSpendingPasswordDialogContainer extends Component<Pro
   static defaultProps = {
     actions: null,
     stores: null,
-    forceSetPassword: false,
   };
 
   render() {
-    const { actions, forceSetPassword } = this.props;
+    const { actions } = this.props;
     const { uiDialogs, wallets, walletSettings } = this.props.stores;
     const dialogData = uiDialogs.dataForActiveDialog;
     const { updateDataForActiveDialog } = actions.dialogs;
@@ -58,7 +56,6 @@ export default class ChangeSpendingPasswordDialogContainer extends Component<Pro
         }}
         isSubmitting={updateSpendingPasswordRequest.isExecuting}
         error={updateSpendingPasswordRequest.error}
-        forceSetPassword={forceSetPassword || false}
         walletName={activeWallet.name}
       />
     );