+++ storybook/stories/wallets/_utils/defaultWalletProps.js
@@ -62,7 +62,6 @@ export const defaultProps = {
   changeSpendingPasswordDialog: (
     <ChangeSpendingPasswordDialog
       walletName={text('Wallet Name', 'Wallet Name')}
-      forceSetPassword={false}
       currentPasswordValue="current"
       newPasswordValue="new"
       repeatedPasswordValue="new"